#1681 built beam lib with emcc

+++ 3rdparty/libbitcoin/src/unicode/unicode_streambuf.cpp
@@ -36,7 +36,7 @@ unicode_streambuf::unicode_streambuf(std::wstreambuf* wide_buffer, size_t size)
     narrow_(new char[narrow_size_]), wide_(new wchar_t[narrow_size_]),
     wide_buffer_(wide_buffer)
 {
-    if (wide_size_ > (bc::max_uint64 / utf8_max_character_size))
+    if ((uint64_t)wide_size_ > (bc::max_uint64 / utf8_max_character_size))
         throw std::ios_base::failure(
             "Wide buffer must be no more than one fourth of max uint64.");
